daq:Options Remove unnecessary explicit user-defined conversion operators.
Clang says:
{{{
/.../trunk/utilities/options/include/Options.h:286:2: error: conversion function converting 'OptionMap<std::string>' to its base class 'std::map<std::basic_string<char>, std::basic_string<char>, std::less<std::basic_string<char> >, std::allocator<std::pair<const std::basic_string<char>, std::basic_string<char> > > >' will never be used [-Werror]

        operator const std::map<std::string, std::string>& () const {
        ^
1 error generated.
}}}
at least for explora-light, since DAQ does not compile with clang due to invalid syntax in commoncpp headers.

Indeed, this is not needed: Conversions to base-class are preferred over user-defined conversion operators with same signature so this is dead code.
Interestingly, clang did not say that for the unspecialized template, but I also removed it from there.

git-svn-id: file:///net/homes/repositories/daq/trunk/libs/options@5731 d3e46cb6-f2d9-11dd-84a8-d10596e79348