cmake: Remove attempt to use native cmake-packaging again.
It requires recent cmake (>2.8.8 at least).
That's seldomly available on the platforms we want to package for.
Also, the user of these native macros would need a recent cmake,
since we EXPORT targets and they have to IMPORT:
https://cmake.org/cmake/help/git-master/manual/cmake-packages.7.html#creating-packages
Advantage of this technique would be that cmake allows to
event include this project inside the user's build-tree as an external,
without any changes.
However, since it's basically nowhere available and complicated,
the easiest solution should be to just provide a simple FindOptionParser.cmake
which relies on pkgconfig. We don't support pkg-config-less stuff
like Windows anyways.