Help: Add line-wrapping for long explanations.
For now, a hardcoded maximum explanation-length
(before wrapping occurs) of 100 is used.
We try to break at spaces or dashes.
If impossible, we break hard and add a dash
to indicate that.
All broken lines are indented a bit,
and whitespace is not kept when breaking
while e.g. dashes are of course.

The optionExample has been extended
with several examples to illustrate.