Options.h: Remove macro hackery again.
Mingw64 does not offer basename or readlink replacements.
The runtime only has an implementation for the tool "basename"
which seems not exposed.
There's only the not-useful-since-broken-by-design POSIX-basename.
The only solution I see is to use e.g.
http://permalink.gmane.org/gmane.comp.gnu.mingw.devel/1268
as a solution in case we do not have basename from glibc,
since the basename-code attached there seems solid and
takes Windows specials into account.

There's still no easy solution for "readlink", though.