Make exit behavoiur configurable.
Now there is a new function fComplainAndLeave that outputs help (if
the optional parameter is true (default)) and exits. In a class
derived from OptionParser this can be overloaded to e.g. throw an
exception.
Also the complaints are not necessarily written to std::cerr but a
configurable output stream.