CMake: Split build and installation of docs.
"doc" is now part of the "all"-target if the user
specified "BUILD_DOCUMENTATION".
That's a better interpretation of that flag.
Also, before, build failed: We tried to install
documentation always, but the build-target was optional...

There's now also a "INSTALL_DOCUMENTATION" option.
Packaging usually wants to keep documentation separate
and may wish to install it on its own.