CMake: Reconsider the doc-generation.
It's even nicer if we always add the "doc" target
if doxygen is there, and just issue a note if doxygen is not there.
Installation is optional now.
Documentation only made if the user explicitly makes the "doc" target.
I.e. for a full build you now need the incantation:
- make
- make doc
- make install
This seems to be the most wide-spread solution;
if doxygen is missing and the packager does not see the warning,
"make doc" will fail as expected.