change source storage system
now the source of options is stored not in strings but sourceItems
that refer to a sourceFile that remembers the file (or commandline)
This way we save some memory as each file name is stored only once
and are more type safe.
Also this will allow later to keep read config file options when
writing confiog files with included files.